+++ package.json
@@ -59,6 +59,7 @@
     "globals": "^15.12.0",
     "jsdom": "^25.0.1",
     "msw": "^2.7.0",
+    "msw-storybook-addon": "^2.0.4",
     "postcss": "^8.4.49",
     "prettier": "^3.4.2",
     "prettier-plugin-tailwindcss": "^0.6.9",
