+++ src/page/LoginPage/components/LinkButton/LinkButton.stories.tsx
@@ -1,18 +1,29 @@
 import { Meta, StoryObj } from '@storybook/react';
-import { MemoryRouter, Routes, Route, useLocation } from 'react-router-dom';
+import { Routes, Route, useLocation } from 'react-router-dom';
 import LinkButton from './LinkButton';
 
 const meta: Meta<typeof LinkButton> = {
   title: 'page/LoginPage/Components/LinkButton',
   component: LinkButton,
   tags: ['autodocs'],
+  decorators: [
+    (Story) => (
+      <>
+        <Routes>
+          <Route path="/" element={<div>Login Page</div>} />
+          <Route path="/table" element={<div>Table List Page</div>} />
+        </Routes>
+        <Story />
+        <CurrentRoute />
+      </>
+    ),
+  ],
 };
 
 export default meta;
 
 type Story = StoryObj<typeof LinkButton>;
 
-// 라우트 전환 결과를 표시하는 컴포넌트
 const CurrentRoute = () => {
   const location = useLocation();
   return <div>현재 경로: {location.pathname}</div>;
@@ -23,17 +34,4 @@ export const Default: Story = {
     url: '/table',
     title: '로그인',
   },
-
-  decorators: [
-    (Story) => (
-      <MemoryRouter initialEntries={['/']}>
-        <Routes>
-          <Route path="/" element={<div>Login Page</div>} />
-          <Route path="/table" element={<div>Table List Page</div>} />
-        </Routes>
-        <Story />
-        <CurrentRoute />
-      </MemoryRouter>
-    ),
-  ],
 };
