+++ src/page/TableListPage/components/Modal/DropdownForDebateType.stories.tsx
@@ -1,18 +1,10 @@
 import { Meta, StoryObj } from '@storybook/react';
 import DropdownForDebateType from './DropdownForDebateType';
-import { GlobalPortal } from '../../../../util/GlobalPortal';
 
 const meta: Meta<typeof DropdownForDebateType> = {
   title: 'page/TableListPage/Components/DropdownForDebateType',
   component: DropdownForDebateType,
   tags: ['autodocs'],
-  decorators: [
-    (Story) => (
-      <GlobalPortal.Provider>
-        <Story />
-      </GlobalPortal.Provider>
-    ),
-  ],
 };
 
 export default meta;
