+++ src/page/TableListPage/components/Table/Table.stories.tsx
@@ -1,11 +1,25 @@
 import { Meta, StoryObj } from '@storybook/react';
-import { MemoryRouter, Routes, Route, useLocation } from 'react-router-dom';
+import { Routes, Route, useLocation } from 'react-router-dom';
 import Table from './Table';
 
 const meta: Meta<typeof Table> = {
   title: 'page/TableListPage/Components/Table',
   component: Table,
   tags: ['autodocs'],
+  decorators: [
+    (Story) => (
+      <>
+        <Routes>
+          <Route path="/table" element={<div>Table List Page</div>} />
+          <Route path="/" element={<div>Table Setup Page</div>} />
+        </Routes>
+        <div className="flex flex-col items-center gap-4 p-4">
+          <Story />
+          <CurrentRoute />
+        </div>
+      </>
+    ),
+  ],
 };
 
 export default meta;
@@ -23,18 +37,4 @@ export const Default: Story = {
     type: '의회식 토론',
     duration: 30,
   },
-  decorators: [
-    (Story) => (
-      <MemoryRouter initialEntries={['/table']}>
-        <Routes>
-          <Route path="/table" element={<div>Table List Page</div>} />
-          <Route path="/" element={<div>Table Setup Page</div>} />
-        </Routes>
-        <div className="flex flex-col items-center gap-4 p-4">
-          <Story />
-          <CurrentRoute />
-        </div>
-      </MemoryRouter>
-    ),
-  ],
 };
