+++ src/page/TableOverviewPage/TableOverview.stories.tsx
@@ -1,45 +1,33 @@
 import { Meta, StoryObj } from '@storybook/react';
 import TableOverview from './TableOverview';
-import { MemoryRouter, Route, Routes } from 'react-router-dom';
 import { DebateInfo } from '../../type/type';
 
+// 1) 메타 설정
 const meta: Meta<typeof TableOverview> = {
   title: 'page/TableOverview',
   component: TableOverview,
   tags: ['autodocs'],
-  decorators: [
-    (Story) => (
-      <MemoryRouter
-        initialEntries={[
-          {
-            pathname: '/overview',
-            state: [
-              {
-                stance: 'PROS',
-                type: 'OPENING',
-                time: 150,
-                speakerNumber: 1,
-              },
-              {
-                stance: 'CONS',
-                type: 'REBUTTAL',
-                time: 180,
-                speakerNumber: 2,
-              },
-            ] as DebateInfo[],
-          },
-        ]}
-      >
-        <Routes>
-          <Route path="/overview" element={<Story />} />
-        </Routes>
-      </MemoryRouter>
-    ),
-  ],
 };
 
 export default meta;
-
 type Story = StoryObj<typeof TableOverview>;
 
-export const Default: Story = {};
+// 2) 스토리 정의
+export const Default: Story = {
+  parameters: {
+    routeState: [
+      {
+        stance: 'PROS',
+        type: 'OPENING',
+        time: 150,
+        speakerNumber: 1,
+      },
+      {
+        stance: 'CONS',
+        type: 'REBUTTAL',
+        time: 180,
+        speakerNumber: 2,
+      },
+    ] as DebateInfo[],
+  },
+};
