+++ src/page/TableSetupPage/TableSetup.stories.tsx
@@ -1,18 +1,10 @@
 import { Meta, StoryObj } from '@storybook/react';
 import TableSetup from './TableSetup';
-import { GlobalPortal } from '../../util/GlobalPortal';
 
 const meta: Meta<typeof TableSetup> = {
   title: 'page/TableSetup',
   component: TableSetup,
   tags: ['autodocs'],
-  decorators: [
-    (Story) => (
-      <GlobalPortal.Provider>
-        <Story />
-      </GlobalPortal.Provider>
-    ),
-  ],
 };
 
 export default meta;
