+++ src/page/TableSetupPage/components/EditDeleteButtons/EditDeleteButtons.stories.tsx
@@ -1,24 +1,11 @@
 import { Meta, StoryObj } from '@storybook/react';
 import EditDeleteButtons from './EditDeleteButtons';
 import { DebateInfo } from '../../../../type/type';
-import { MemoryRouter } from 'react-router-dom';
-import { GlobalPortal } from '../../../../util/GlobalPortal';
 
 const meta: Meta<typeof EditDeleteButtons> = {
   title: 'page/TableSetup/components/EditDeleteButtons',
   component: EditDeleteButtons,
   tags: ['autodocs'],
-  decorators: [
-    (Story) => (
-      <MemoryRouter>
-        <GlobalPortal.Provider>
-          <div className="h-screen">
-            <Story />
-          </div>
-        </GlobalPortal.Provider>
-      </MemoryRouter>
-    ),
-  ],
 };
 
 export default meta;
